Comment Toggling, Defined Parameters, Minor Fixes and Changes.
Comment toggling via keyboard shortcut now works for .ks files.  This feature is available through the kos.tmlPreferences file, which should be added to User directory with the .tmlLanguage file.

Single parameter declarations (one definition on a line) now receive highlighting like paramater declarations in Python.  This feature may be reverted to cut down on highlighting in general.  Trying to highlight multiple definitions on the same line is still being worked on.  The current issue comes down to getting the regex search to keep matching after the first match is found.  I think the issue has to do with the postive lookbehind I'm using.  The only workaround is just highlighting everything found after the word parameter, but that introduces too much color.

Smaller updates include the following:
- @lazyglobal is now highlighted like function delegates
- Fixed a bug where bound variables were still being highlighted incorrectly when those variables were being used as a parameter.  This error was the result of the code restructuting push a few commits back.

README was updated with new information.  Some of the README was restructured to accomodate for new updates, improve readability, and to make some sections more concise.