Project File/Directory Restructure:
The YAML-tmLanguage file, along with the OriginalConver (now named BaseConversion) was moved to a new directory, src.  This was done to seperate the source code used to create the final syntax definition from the definition itself. README was updated to reflect these changes, and image links were updated so they would be based off what's on the main develop branch