Convert build system to Gradle
This roughly follows the [IntelliJ plugin template](https://github.com/JetBrains/intellij-platform-plugin-template/blob/master/build.gradle.kts) to convert the project to be built with Gradle. This should allow folks to compile the project more easily and reliably, and is also the recommended way to build projects moving forward.

As part of this, and following [best practices (1.3)](https://github.com/JetBrains/Grammar-Kit/blob/master/HOWTO.md#i-general-notes) and moves the generated sources into a separate source root, in this case the `gen` folder.

Note that the main difference between this project and the default/recommended/template project is the source layout. Namely, they expect `src` dir that has `java`, `resources`, etc in it, whereas our source code is directly in the `src` directory. If folks here are open to it, I think moving the code could make it easier to work with the plugin moving forward.