[idea] Clean up generics, add basic tests
I believe that the generic parameters that were on a number of classes like `Duality` and `ContextBuilder` that were incorrectly added. In particular, the generics were being specified for only a small percentage of the usages of the classes themselves, and even those often eventually ended up using a raw generic (without parameters) which turns off all generic type checks anyway. I was trying to remove all usages of raw generics, and found some of the types to be impossible to set to anything valid. For instance, the `InstructionsParser` takes a `ContextBuilder`, but we can't set a `Flow` type on that builder since the `InstructionsParser` adds heterogeneous types to it. Similarly, we have a `resolve` method which ends up returning a lookup out of a heterogeneous map, and there's no way at compile time (eg, using generics) to know that type returned from this map is what we want. What we actually need in order to be safe in these cases is runtime type casts. Those few places that do need type casts have been added, and most of the other usages of these classes remain the same even after the generics have been removed.

To enforce that we don't use raw generics in the future, this turns on `javac`'s lint warnings and has them be compile-time errors. Also while going through these files, I took IntelliJ's suggestions for fixes/cleanups.

I still don't have a great handle on all of this flow/differentiation code, so it's quite possible I'm missing a core concept that this breaks. Similarly, it's still unclear to me whether things like `Flow` actually need to be generic.

This also adds some basic tests for some functionality (really just copying the examples from the `README.md`) so I could quickly and easily verify I hadn't blatantly broken things. Some of the test framework is copied from analogous tests and support files in [the `intellij-rust` plugin](https://github.com/intellij-rust/intellij-rust/tree/master/src/test/kotlin/org/rust). It adds a [GitHub Action](https://docs.github.com/en/actions) to build the plugin and run the tests for it on PRs and commits that touch the `IDEA/` folder.