[idea] Convert some files to Kotlin
I've been working on adding an annotator to point out missing/incorrect variables, and in doing so ended up converting a number of files to Kotlin for easier modification. In order to separate the syntactic changes due to converting to Kotlin from the semantic changes to in my annotator diff (especially since converting to Kotlin usually prevents Git from seeing it as a rename), I've pulled the conversion out to a separate PR.

This also adds a few nullity annotations while here to try to make it a bit clearer what can and cannot be null (since Kotlin cares about that).