[idea] Add an "unknown identifier" annotator, improve resolving
This adds an annotator that marks any named identifier that can't be resolved to its original declaration with an error. At the same time, this improves resolving identifiers to handle a few extra situations:

- Resolve globals and imports (which must be globals) in any scope in the file (maybe questionable if we actually want this)
- Imported files only contribute globals, since each file has its own local scope
- Ensure that locals are declared before used; works when declaration is in a parent scope and handles the usage being inside a function
- Correctly set global and local occurrence types on declarations
- Resolve `LOCK`, `FROM`, and `FOR` declarations
- Resolve builtin globals
- Allow functions to look like variables (both at declaration and use time)

Note that the annotator still has a number of false positives (eg marks things that are actually correct as errors) so it's disabled outside of tests, but this allows it to be incrementally developed. Some areas that still need doing:

- Handling `@LAZYGLOBAL`; right now it only looks for explicit declarations, so a cheap/easy way to "handle" it up front would be to only turn the annotator on for `@LAZYGLOBAL OFF`
- Resolving globals from `RUN ...`/`RUN ONCE ...` called with string literals; while this likely isn't a blocker, it does allow resolving imports inside directories which would be nice

This also adds a few more nullity annotations in areas I ran across, to help me (and Kotlin) keep track of what can/cannot be null.