[idea] Implement parser recovery
This implements recovery in the parser such that partially complete statements don't completely break the PSI. This is going to be useful for auto-complete so you could write, eg `PRINT ` (without at `.`) and still allow the parser to understand the rest of the file.

This is based off of the documentation at the Grammar Kit [README](https://github.com/JetBrains/Grammar-Kit#attributes-for-error-recovery-and-reporting) and [HOWTO](https://github.com/JetBrains/Grammar-Kit/blob/master/HOWTO.md#22-using-recoverwhile-attribute), the [`intellij-rust` grammar](https://github.com/intellij-rust/intellij-rust/blob/master/src/main/grammars/RustParser.bnf#L306-L335), and some trial-and-error. It was particularly tricky to both recover at a keyword (interpreting it as the beginning of the next instruction) and allow keywords to be used as variables (which is apparently something you can do in KerboScript?!?). To do this, we rely on an `ident` rule which combines `IDENTITY` and the various keywords and remove the lexer type remapping, which prevented the recovery from working (since things that looked like `IDENTITY`s would be auto-converted to such). This does add one more layer to the PSI, but that seems like a worthwhile tradeoff.

Most of the rest of the changes are handling this PSI change, looking for the new `KerboScriptIdent` PSI element (or the file version) rather than the `IDENTITY` (or file version) lexer token. To take advantage of nice functions like `firstOrNull` I've also converted a few files to Kotlin. The files should be renamed in a separate commit to make viewing the actual changes to the file content easier.