[idea] Fix overwriting of LocalContext cache
It turns out, as prompted by the Annotator failures in https://github.com/KSP-KOS/EditorTools/pull/66#issuecomment-701490170, there was a bug in how we on-demand construct context's for files. If the first time we interact with a file's contexts is by asking for the scope of some element within the file, that would trigger the entire file to be walked, which would itself create contexts for all scopes. However, once that finished the initial call would overwrite the `LocalContext` that had been created for the scope causing lookups on that scope to fail. This avoids that issue by using a cache that was created on the parent scope, if any.

Note that the reason we don't get into an infinite loop is because the `KerboScriptFile` checks if it's already in progress and early-aborts if so:
https://github.com/KSP-KOS/EditorTools/blob/844afcbdc7f0d7e211319c43af05d697855a1df2/IDEA/src/main/java/ksp/kos/ideaplugin/KerboScriptFile.java#L73

It's unclear how to fashion a test to trigger this failure. It seems that the annotator test failed in https://github.com/KSP-KOS/EditorTools/pull/66 "due to" upgrading to IDEA 2020.2 (eg some internal ordering of when/how it uses files likely changed).